fix: auto commit to DB

+++ model/trashcart_model.go
@@ -9,8 +9,8 @@ type Cart struct {
 	UserID              string     `gorm:"not null" json:"user_id"`
 	User                User       `gorm:"foreignKey:UserID;constraint:OnDelete:CASCADE;" json:"-"`
 	CartItems           []CartItem `gorm:"foreignKey:CartID;constraint:OnDelete:CASCADE;" json:"cart_items"`
-	TotalAmount         float32    `json:"total_amount"`
-	EstimatedTotalPrice float32    `json:"estimated_total_price"`
+	TotalAmount         float64    `json:"total_amount"`
+	EstimatedTotalPrice float64    `json:"estimated_total_price"`
 	CreatedAt           time.Time  `gorm:"autoCreateTime" json:"created_at"`
 	UpdatedAt           time.Time  `gorm:"autoUpdateTime" json:"updated_at"`
 }
@@ -21,8 +21,8 @@ type CartItem struct {
 	Cart                   *Cart         `gorm:"foreignKey:CartID;constraint:OnDelete:CASCADE;" json:"-"`
 	TrashCategoryID        string        `gorm:"not null" json:"trash_id"`
 	TrashCategory          TrashCategory `gorm:"foreignKey:TrashCategoryID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;" json:"trash_category"`
-	Amount                 float32       `json:"amount"`
-	SubTotalEstimatedPrice float32       `json:"subtotal_estimated_price"`
+	Amount                 float64       `json:"amount"`
+	SubTotalEstimatedPrice float64       `json:"subtotal_estimated_price"`
 	CreatedAt              time.Time     `gorm:"autoCreateTime" json:"created_at"`
 	UpdatedAt              time.Time     `gorm:"autoUpdateTime" json:"updated_at"`
 }
